pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.16"
langchain-community = "^0.2.16"
langchain-openai = "^0.1.23"
langgraph-checkpoint-sqlite = "^1.0.3"
beautifulsoup4 = "^4.12.3"
langchain-chroma = "^0.1.3"

[tool.poetry.group.dev.dependencies]
setuptools = "^74.1.1"
gradio = "^4.42.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
